Is it non-deterministic interference between the two tests?
